+++ app/build.gradle
@@ -4,12 +4,11 @@ plugins {
     id 'kotlin-kapt'
     id 'com.google.dagger.hilt.android'
     id 'kotlin-parcelize'
-    id("com.google.gms.google-services")
 }
 
 android {
     namespace 'com.example.moodyapp'
-    compileSdk 34
+    compileSdk 33
 
     defaultConfig {
         applicationId "com.example.moodyapp"
@@ -60,8 +59,7 @@ dependencies {
     implementation 'androidx.compose.ui:ui'
     implementation 'androidx.compose.ui:ui-graphics'
     implementation 'androidx.compose.ui:ui-tooling-preview'
-    implementation 'androidx.constraintlayout:constraintlayout-compose:1.0.1'
-    implementation 'androidx.compose.material3:material3-android:1.2.1'
+    implementation 'androidx.compose.material3:material3'
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
@@ -69,7 +67,6 @@ dependencies {
     androidTestImplementation 'androidx.compose.ui:ui-test-junit4'
     debugImplementation 'androidx.compose.ui:ui-tooling'
     debugImplementation 'androidx.compose.ui:ui-test-manifest'
-    implementation "androidx.compose.material:material-icons-extended:1.6.6"
 
     //Splash Api
     implementation "androidx.core:core-splashscreen:1.0.1"
@@ -111,15 +108,5 @@ dependencies {
     kapt "androidx.room:room-compiler:$room_version"
     implementation "androidx.room:room-ktx:2.5.2"
 
-    //Firebase
-    implementation(platform("com.google.firebase:firebase-bom:32.8.1"))
-    //Analitycs
-    implementation("com.google.firebase:firebase-analytics")
-    //Authentication
-    implementation 'com.firebaseui:firebase-ui-auth:7.2.0'
-    // FirebaseUI for Firebase Auth
-    implementation 'com.firebaseui:firebase-ui-auth:8.0.2'
-    // FirebaseUI for Firebase Realtime Database
-    implementation 'com.firebaseui:firebase-ui-database:8.0.2'
 
 }
