--- app/google-services.json
@@ -1,29 +1,0 @@
-{
-  "project_info": {
-    "project_number": "1022743550481",
-    "project_id": "moodyapp-717a5",
-    "storage_bucket": "moodyapp-717a5.appspot.com"
-  },
-  "client": [
-    {
-      "client_info": {
-        "mobilesdk_app_id": "1:1022743550481:android:3be3ed5bd0cbbd43ef2fb0",
-        "android_client_info": {
-          "package_name": "com.example.moodyapp"
-        }
-      },
-      "oauth_client": [],
-      "api_key": [
-        {
-          "current_key": "AIzaSyByUJJGmJYK6Qq_LCUymhSTI7zTYvmFuT4"
-        }
-      ],
-      "services": {
-        "appinvite_service": {
-          "other_platform_oauth_client": []
-        }
-      }
-    }
-  ],
-  "configuration_version": "1"
-}
