+++ app/src/main/AndroidManifest.xml
@@ -10,24 +10,19 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.MoodyApp"
-        tools:targetApi="34">
+        android:theme="@style/App.Starting.Theme"
+        tools:targetApi="31">
         <activity
-            android:name=".presentation.login.Login"
+            android:name=".MainActivity"
             android:exported="true"
-            android:label="@string/title_activity_login"
-            android:theme="@style/Theme.MoodyApp">
+            android:label="@string/app_name"
+            android:theme="@style/App.Starting.Theme">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity
-            android:name=".MainActivity"
-            android:exported="false"
-            android:label="@string/app_name"
-            android:theme="@style/Theme.MoodyApp"></activity>
     </application>
 
 </manifest>
