+++ app/src/main/res/values/strings.xml
@@ -1,18 +1,3 @@
 <resources>
     <string name="app_name">MoodyApp</string>
-    <string name="loginButton">Login</string>
-    <string name="title_activity_login">LoginActivity</string>
-    <string name="greetings">Hello</string>
-    <string name="user">Username</string>
-    <string name="pass">Password</string>
-    <string name="subtitleLogin">You must be login to continue</string>
-    <string name="forgotPass">Did you forget your password?</string>
-    <string name="noAccount">You do not have an account?</string>
-    <string name="registerHere">Sign up here</string>
-    <string name="SignInButtonLabel">Sign in</string>
-    <string name="googleSignin">Sign in with Google</string>
-    <string name="TitleAlertComplete">Error</string>
-    <string name="ContentAlertComplete">All fields must be complete to continue</string>
-    <string name="ConfirmMsg">Confirm</string>
-    <string name="NotLoginAlertContent">The username or password is incorrect</string>
 </resources>
