+++ app/src/main/res/values/themes.xml
@@ -2,5 +2,4 @@
 <resources>
 
     <style name="Theme.MoodyApp" parent="android:Theme.Material.Light.NoActionBar" />
-
 </resources>
